Fix: fixed the map names + moved some styling
