+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styles from "./page.module.css";
-import List from "../components/List/list";
+import List from "../components/List/List";
 
 
 async function getData() {
