+++ src/components/List/list.module.css
@@ -5,21 +5,6 @@
     
 }
 
-.border{
-    border-top: 1px solid var(--border-color);
-    border-bottom: 1px solid var(--border-color);
-    padding: 20px;
-    border-radius: 10px;
-    margin: 5px;
-}
-
-.border:hover{
-    background-color: var(--background-color);
-    transform: scale(1.01); 
-    transition: transform 0.3s ease;
-}
-
-
 .emptyState{
     color: gray;
 }
