+++ src/components/Todoitem/todoitem.module.css
@@ -1,0 +1,13 @@
+.border{
+    border-top: 1px solid var(--border-color);
+    border-bottom: 1px solid var(--border-color);
+    padding: 20px;
+    border-radius: 10px;
+    margin: 5px;
+}
+
+.border:hover{
+    background-color: var(--background-color);
+    transform: scale(1.01); 
+    transition: transform 0.3s ease;
+}
