+++ src/components/Todoitem/todoitem.tsx
@@ -1,34 +1,17 @@
 import React from "react";
-import Styles from "./list.module.css";
-import { TodoItem } from "../../types/type";
+import Styles from "./todoitem.module.css";
 import { Checkbox } from "../Checkbox/checkbox";
 
+
+
 interface TodoItemProps {
-    item: TodoItem;
-    onCheckboxClick: (id: string) => void;
-  }
+  item: ITodoItem;
+}
 
-
-//   export default function Todoitem({ item }: TodoItemProps) {
-//     // Handle checkbox click to toggle completion
-  
-//     //This is the Empty state. If the array that i get from the Api is empty i return a paragraph
-//     if (item.length === 0) {
-//       return (
-//         <div className={Styles.emptyState}>
-//           <p>Create your first todo</p>
-//         </div>
-//       );
-//     }
-  
-//     return (
-//       <ul className={Styles.ul}>
-//         {data.map((item) => (
-//           <li className={Styles.border} key={item.id}>
-//             <Checkbox label={item.title} />
-//           </li>
-//         ))}
-//       </ul>
-//     );
-//   }
-  
+export const TodoItem = ({ item }: TodoItemProps) => {
+  return (
+        <li className={Styles.border} key={item.id}>
+          <Checkbox label={item.title} />
+        </li>
+  );
+};
